+++ practice_with_YumemiWeather/Data/Request.swift
@@ -1,0 +1,14 @@
+//
+//  Model.swift
+//  practice_with_YumemiWeather
+//
+//  Created by Kai Tanobe on 2021/10/06.
+//
+
+import Foundation
+import UIKit
+
+struct Request: Codable {
+    var area: String
+    var date: Date
+}
