+++ practice_with_YumemiWeather/Data/WeatherState.swift
@@ -1,0 +1,14 @@
+//
+//  WeatherState.swift
+//  practice_with_YumemiWeather
+//
+//  Created by Kai Tanobe on 2021/11/02.
+//
+
+import Foundation
+
+enum WeatherState: String {
+    case sunny = "sunny"
+    case rainy = "rainy"
+    case cloudy = "cloudy"
+}
